GP131: Update TEE_InitRefAttribute()
Update TEE_InitRefAttribute() according to TEE Internal Core API version
1.3.1. The size parameter is changed to use size_t instead of uint32_t.

+++ lib/libutee/include/tee_internal_api.h
@@ -189,7 +189,7 @@ TEE_Result __GP11_TEE_PopulateTransientObject(TEE_ObjectHandle object,
 					      uint32_t attrCount);
 
 void TEE_InitRefAttribute(TEE_Attribute *attr, uint32_t attributeID,
-			  const void *buffer, uint32_t length);
+			  const void *buffer, size_t length);
 void __GP11_TEE_InitRefAttribute(__GP11_TEE_Attribute *attr,
 				 uint32_t attributeID,
 				 const void *buffer, uint32_t length);
